video: msm: somc_panel: Restructure for modularity and add In-Cell.
Restructure the somc_panel driver for modularity: now it
supports modular power sequence (and reset) functions.

This change was needed to add the new In-Cell power interface
driver for new panel+ts combo found in new platforms, such
as Loire and Tone.

Also, we have now splitted legacy panels and In-Cell for
increasing overall driver human readability and decreasing
footprint.