clk: msm: mdss: 20nm-pll: Fix code and enable 20nm pll for MSM8994
In this 3.10 snapshot, we were using readl_poll_timeout_noirq
which was an early API which was not accepted on mainline.
On the current (3.18) kernel, the API got refined and
accepted, but the call is not named readl_poll_timeout_atomic.

Now that the 20nm pll driver can build, enable it in Makefile.