bluetooth: bluesleep: Control serial clocks from rfkill block
If we block the rfkill switch, then we are not using bluetooth
for sure: disable also the serial clocks in this case and
enable them when unblocking the rfkill switch.
Also, in our case, the rfkill block function is basically
powering on and off the bluetooth chip via GPIO setting.