arm: DT: Suzuran: Add platform regulator/lane settings to MDSS
We have platform regulator and lane settings as SoMC properties
in specific dsi panel configuration.
This is not really working fine in kernel 3.18, and then we
anyway have no different regulator or lane setting for each
panel.

The best way is to go the qcom way, defining the settings in
dsi node directly, instead of waiting for panel probe and
reading it as proprietary setting.