media: msm: msm_v4l2_vidc: Add legacy VIDC platform version probing
New 3.18 devices have a qcom,platform-version property which
is specifying the version mask and shift.
Old devices have no property and this is required for platform
version calculation.

Adding a property will be harder: value is u32, so, add the
fallback mechanism. The values in there are valid for ALL
old SoCs (8974, 8994, 8939, 8916, 8936 etc).