video: msm: somc_panel: Fix GPIO direction handling in main driver
New pinctrl from mainline does not set the GPIO direction
automatically, by supposing that if we want to enable a GPIO
we want it to be output.
This is VERY good, but our driver was made to work on QC TLMM
pinctrl, which was doing the opposite.

Fix the GPIO direction handling in main somc_panel panel_driver
in functions that are common to both legacy and incell panels.