fbdev: somc_panel: Support dynamic setting of the escape clock
In kernel 4.4, there's no more default setting of the DSI host's
escape clock. This clock is static in _almost_ all cases for QC
and the usual default for ALL QC SoC at the time of writing this
(the newest one right now is MSM8998) is 19.2MHz.

Produce a default setting of 19.2MHz for the escape clock and
support overriding it with a DT property, as per QC spec.