arm: DT: MSM8956: Rewrite VIDC IOMMU and scaling configurations
In commit 6622201e1ade I have introduced a new IOMMU configuration
that is using the arm-smmu driver with the new qualcomm code.
As referenced in that commit, the context banks are now deprecated
and we need to also rewrite the IOMMU configuration in all of
the drivers' DT nodes.

Rewrite the IOMMU configurations for the VIDC driver to take the
right IOMMU SIDs on the APPS dedicated IOMMU, as initialized by
the bootchain in TrustZone.
Also, specify the virtual address pools that we are allowed to
read and write into as these are being also configured in the
IOMMU by the secure bootchain.