iommu: qcom: iommu-v1: Handle secure IOMMU correctly
We cannot halt, nor program the context, assign ASIDs etc on
the secure part of the IOMMU. This will lead to a very bad
hardware fault with subsequent reboot with apparently no reason.

This happens because of a security feature set in the early
bootchain and, when the secure bit is set, it is NOT possible to
reprogram the IOMMU contexts in ANY way.