arm: DT: Yoshino: Start clearpad right after driver probe
This is necessary to not include userspace hacks for it.
Originally, the property is zero because this can cause problems
in conditions where the firmware has to be flashed to the device.

On AOSP, we don't support clearpad firmware upgrade, hence we
can just enable the touchscreen without userspace assistance.