fbdev: msm: somc_panel: Send off commands, set VFP on secondary panel
When we have a dual-panel or a split-panel with dual control
configuration, we don't necessarily need to send the off command
to only one panel or one half.
Moreover and more importantly, if we have sone dynamic FPS data
to update at poweroff time, we want to do it for both right and
left half of the display, otherwise we may get different FPS for
one of the two controls, producing very bad tearing problems.