arm: DT: MSM8956: Configure fixedsids adsp fastrpc context banks
MSM8916-class SoCs, including MSM8956, have the old MMU-500 IOMMU
that is using fixed context banks, preallocated by TZ, and
meant for an old concept of the Hexagon MMU compute banks.
Now that the adsprpc driver supports compatibility with it,
declare the fixedsids compute context banks with the correct
subsids and address pool.