arm64: Kconfig: MSM8916: Declare HAVE_FUTEX_CMPXCHG
ARM64 has got a working implementation of the required function
futex_atomic_cmpxchg_inatomic(), which is proven to be always
working for the MSM8916-class SoCs.
This should be also true for another number of Qualcomm SoCs,
but better safe than sorry, enable only here.
This will avoid init checks on futex driver and strange crashes
on at least this SoC.