arm: DT: msm8956-regulator: Fix voltage constraints for APC0 and GFX
The APC0 (pm8950_s5) and GFX (pm8004_s5) regulators constraints
were wrong and never respected, due to the limitations put in
the HW (by the bootloader?).
The HW constraints are currently configured to be higher than
what it was declared as minimum in DT, so that minimum voltage
was never getting set.

Fix the constraints in DT so that we get the right voltages set
whenever requested.