arm: DT: msm8956-gpu: Fix kgsl-3d0 clocks and stabilize GPU
The reference DT from 1.3.3 release was misleading and wrong:
we had a GTCU AHB clock declared in the kgsl-3d0 clocks that
was then never getting used by the kgsl pwrctrl driver.
On recent driver upgrades, this clock started getting used
and this is one of the things that started producing issues,
as this clock should never be disabled (gets enabled by
the bootloader and, if we disable it, we may run into private
GPU memory corruption issues).

More than this, the TBU1 clock (specific to MSM8956), shall
be enabled prior to the TBU clock, but only after the
alwayson one otherwise, after some en/disablement cycles,
that (tbu1) clock will get unbalanced and the GPU will
fault.