arm: DT: Dora: Enable USB OTG functionality with new ID detection
The newest version of our driver can now get the ID state through
a VADC channel that is forcefully called "usb_detect".
Change the channel name to reflect that, the instruct DWC3 to get
the state from the proper VADC and relative channel.
Also, signal that Type-C is missing on this board to trigger DWC3
internal ID work.