fbdev: msm: somc_panel: Remove old hack for DriverIC power
This old hack was used for SoMC Rhine and SoMC Yukon: these
platforms are unsupported on the current kernel releases and
they will never come back.
In any case, even if we'll need something like this in the
future, at least on k4.4, there are other ways of doing that
in a much cleaner way, completely inside somc_panel, without
external code.