mmc: host: sdhci-msm: keep clocks enabled if suspend fails
If clocks are found to be enabled while suspending, then the suspend is aborted.
However, it is possible for the clocks to be disabled by whatever caused them to be enabled at the time of suspend.
This leads to a race condition just before the suspend retry is triggered, where the clocks are disabled just as another task is trying to access the controller.

When this happens, the device crashes with the error:
[18096.907961] Unhandled fault: synchronous external abort (0x96000010) at 0xffffff800a93e928

To avoid this, make sure the clocks remain enabled when suspend fails by calling sdhci_msm_enable_controller_clock(host).

Change-Id: I554027c6c7f04d91cd773510c94dbc45f4eaa5f3