clk: qcom: gcc-msm8976: Add vdd_dig_gfx regulator for corner voting
The driver was trying to use the GCC vdd_dig to vote for voltages,
but this was creating race conditions because of the GFX3D driver
being splitted from the GCC one, even if it is in the same file.
Okay, it's not clean and the files will probably get splitted but,
for now, solve the race by declaring a GFX3D vdd_dig regulator
and using it in the GFX3D clock.