arm64: mm: dma-mapping: Don't check domain attached for IOMMUv1/v2
This is a hack for a hack!
Whenever the IOMMU tries to get mapped to a device via dma-mapping,
it will get attached to the DMA API's own domain: this happens
between of_xlate and add_device and it is wrong.

There is no way to actually fix that properly, until the ARM64 DMA
API gets converted to rely on IOMMU groups and domains and to make
things even worse, this file also contains some hacks from qcom
for their custom implementations.

Fix the whole thing by adding an ifdef for the QCOM_IOMMU_V1
driver that we are using for legacy SoC IOMMUs until the ARM64
DMA API gets migrated to the proper "things".