clk: gcc-msm8976: TEMP: Split support for software voter clocks
On the newest clock framework we have no support for voter
clocks inside of the qcom common helpers.
Register a new clock controller to support voter clocks only.

This commit is temporary and will be reverted once proper
support will be added to the qcom specific part of the common
clock framework.