pinctrl: msm8996: Replace with driver from Open Devices k4.4
The driver in k4.9 was not entirely compatible with whatever
we have.
This can probably be revisited with the pinctrl driver from
a later kernel version but, for now, let's just use a tested
working version of the pinctrl driver for msm8996, from the
Open devices kernel branch LA.UM.6.4.r1.