iommu: Import ARM-SMMU driver from Open Devices k4.4 as LEGACY
The ARM-SMMU driver in kernel 4.9 is specially crafted for the
new SDM670 and SDM845 SoCs and it doesn't support old ones such
as MSM8996, MSM8998, SDM630/660.

Import the driver from Open Devices kernel 4.4 LA.UM.6.4.r1 and
name it as a legacy one, until (and if) the main one gets fixed
to work properly with these SoCs.