clk: qcom: alpha-pll: Add ACPU soft FSM voting mechanism
On some SoCs the GPLL0 may be a clocksource for the CPU, but with
a special trick: the CPU requires XO_A as GPLL0 source, while
the other clocks require the regular XO.
To achieve this, the clock controller driver is declaring the
GPLL0 twice: one having parent XO, one having parent XO_A, hence
we may get no users on one of the two devices and getting the
GPLL0 unexpectedly down for all users.
To fix this, add a soft voting mechanism that permits us to
achieve a functionality that is similar to HW FSM, but done in
the software driver.