media: msm: vidc_3x: Fix Kconfig for folder location changes
Being a k4.4 driver, the tree was different there, as this was
in drivers/media/platform/msm/vidc, but now it is in folder
vidc_3x instead.

Do the required changes for that and, while at it, also change
the Kconfig dependencies for VMEM to the correct new vidc
configuration option.