clk: qcom: gcc-8996: Add 8996-v2/v3 support and refactor SDCC clks
The SDCC clocks need to use the newly introduced clk_rcg2_floor_ops,
as we definitely want to get a floor vote instead, otherwise we
get useless clock spikes on the controller.
This is done in a partial refactoring of some clocks, such as
usb30_master, pcie_aux and, of course, all sdcc clocks.

This adds support for MSM8996v2 and v3 (with pro being v3 in this
driver) by adding a fixup function that switches between v1 and
v2 clock plan for the aforementioned clocks when the right
DT compatible property is set.