pci: dwc: Do not compile pcie-hisi unconditionally when ARM64
This is not really working for us on k4.9, since this is a
backport of the whole thing from k4.17 and misses some ACPI
commits.
In any case, we won't use any HiSilicon chip on this kernel,
hence it's totally safe to just forget about it, as ugly
as this is.