arm64: MSM8998: Force de-advertising 16K granule support
This SoC features a Cortex-A53 compatible cluster of Kryo cores
and a Cortex-A73 compatible cluster of Kryo cores.

Cortex-A53 does not support 16KB memory translation granule
size, while A73 does: this produces a warning at cpu features
discovery due to inconsistency in the MMFR0 register between
the CPUs and taints the kernel.

Force de-advertising 16K granule to remedy to this.