clk: qcom: vdd-level-8998: Fix typo for VDD_MX fmax map macros
The VDD_MX related fmax map macros were declaring VDD_DIG_NUM
as num_rate_max: this isn't really wrong, as the number of
voltage levels of VDD_DIG is the same as the VDD_MX, but
if anyone changes them for some reason in the future, the MX
maps will get screwed.
In any case, changing the num_rate_max to VDD_MX_NUM on the
MX related nodes increases sanity and it's just the right
thing to do.