input: misc: qpnp-power-on: Add option to provoke kernel panic
This option enables code that provokes a kernel panic whenever
KPDPWR and RESIN get pressed: this combination normally triggers
a PMIC-driven emergency reboot after a predefined time, but some
devices powered by an EFI bootloader will delete the pstore logs
in RAM for security purposes if certain bits signaling kernel
panic are not set before rebooting.

This is an option used for early kernel porting or some kind of
debugging that would be impossible otherwise.

This option shall be used ONLY temporarily and ONLY FOR DEBUGGING.