clk: qcom: gcc-msm8998: Add bi_tcxo fixed clock with cxo parent
This is a clock that is found on RPMH powered SoCs, such as
SDM845 and SDM670 and it is just the direct equivalent of
our RPM CXO clock.
Add a fixed clock with that name to satisfy the MDSS 10nm PLL
driver VCO clock dependency, as that driver is shared between
MSM8998, SDM670 adn SDM845.